Fix bug in installation
Before I was doing this, following apenwarr's installer:

* Make the "bin" files python files that import the "lib" files.

* E.g., bin/redo-ifchange imports lib/redo_ifchange.py

However, I had never tested apenwarr's installer. On some systems, this
method causes sqlite I/O errors for reasons that I do not understand at
this time.

I'm falling back to this:

* Make the "bin" files shell scripts that invoke the "lib" files
  directly.

* E.g., bin/redo-ifchange invokes lib/redo-ifchange.py

This is similar to how I was using redo before (just invoke the tool
files directly, with everything in the same directory) and it seems to
work.

As a side effect, I don't have to replace dashes with underscores
anywhere, because the tool names with dashes in them are no longer
python modules.