Revise behavior when redo t would clobber a directory
Prior to this commit, if a redo script would clobber an existing
directory, then redo would crash with an unhandled python exception. The
resulting error message was often mysterious, even for experienced redo
users: for example, it would complain that a directory was not empty or
that an unspecified operation was not permitted. I revised the behavior
to print a specific warning and keep going.

I thought about the following alternative behaviors:

1. Just clobber the directory. This seems redo-y at first, because it
allows you to overwrite output directories (referred to as $3) just like
files. However, there is a key difference: while redo knows when you
have created or modified a file outside its purview, it doesn't know
this for directories. For example, you could create a directory foo
containing several files and later, forgetting you had done that, write
a script called foo.do.  Under this approach, 'redo foo' would silently
clobber the existing foo directory. This seems bad. It seems better to
be conservative and always force the user to explicitly delete
directories (this also seems to be what the original author had in
mind).

2. Halt with an error message and nonzero exit status. This is
potentially good because it signifies that something is seriously wrong.
However, it is more difficult to implement. Also, the warning message
seems consistent with the way redo already behaves when it warns and
refuses to clobber files that have been created or modified by hand. So
I went with the easier solution here. I think the most important points
are (1) not to crash and (2) to print helpful diagnostic messages.