Revise rules for clobbering directories
In general, make directories behave like files:

1. Don't clobber a directory we don't know about.

2. Do clobber a directory if we do know about it.

This revises the behavior described in the following commits:

(1) commit f657dea710e1540eeb418bf74ab85148ebfc6cd1
(2) commit 29925fe1ab166df5f3b36150d7dfccb084f4928e

I think it is a better solution to the problem mentioned in commit (1).
In particular, if we manually create a directory foo and then later try
to run foo.do, redo will warn that foo exists and is not generated, and
will not clobber the directory -- just as for a file.

Note that the distinction between files and directories mentioned in the
log for commit (1) (i.e., redo doesn't know about directories in the
same way as it knows about files) seems to be true of the original redo
implementation, but not intrinsic to the problem. I erased this
distinction.